requirements.txt
aiofiles==0.4.0
amqp==2.5.1
celery==4.3.0
dnspython==1.16.0
kombu==4.6.5
pymongo==3.9.0
pytz==2019.2
redis==3.3.8
requests==2.18.4
sanic==19.6.3
sanic-openapi==0.6.0
uvloop==0.13.0
